Mark SwAssert declarations as noinline (#4304)
* Mark SwAssert declarations as noinline

Without this change, a compiler using LTO may inline the zeros that are
inserted as additional parameters for the call to defaultSwAssert. This
increases code size to optimize for the off-nominal condition where an
assertion is tripped.

In my testing, this single change reduces my largest deployment's
code size by 11%.

* Add 'noinline' to dictionary

* Update Fw/Types/Assert.hpp

Co-authored-by: Rob Bocchino <bocchino@icloud.com>

---------

Co-authored-by: M Starch <LeStarch@googlemail.com>
Co-authored-by: Rob Bocchino <bocchino@icloud.com>